requirements.txt
crawl4ai
chromadb
sentence-transformers
openai
gradio
python-dotenv
nest_asyncio
playwright
